add config and example enviroment

+++ db/connection.go
@@ -18,7 +18,7 @@ func AutoMigrate(DB *gorm.DB) {
 	utils.Info("Update MySQL")
 }
 
-func ConnectionMysql(user, password, host, name string, port int, debugMode bool) error {
+func ConnectionMysql(user, password, host, name string, port uint16, debugMode bool) error {
 	dsn := fmt.Sprintf("%s:%s@tcp(%s:%d)/%s?charset=utf8mb4&parseTime=True&loc=Local",
 		user, password, host, port, name)
